LayoutTests/fast/forms/validation-bubble-escape-key-dismiss-expected.txt

Required text input: 
Tests that the HTML form validation bubble gets dismissed when pressing the Escape key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS validationBubbleContents.message is "Fill out this field"
PASS validationBubbleContents.message is ""
PASS successfullyParsed is true

TEST COMPLETE
